+++ Source/Shaders/Builtin/Constants/degreesPerRadian.glsl
@@ -1,0 +1,16 @@
+/**
+ * A built-in GLSL floating-point constant for converting radians to degrees.
+ *
+ * @alias czm_degreesPerRadian
+ * @glslConstant
+ *
+ * @see CesiumMath.DEGREES_PER_RADIAN
+ *
+ * @example
+ * // GLSL declaration
+ * const float czm_degreesPerRadian = ...;
+ *
+ * // Example
+ * float deg = czm_degreesPerRadian * rad;
+ */
+const float czm_degreesPerRadian = 57.29577951308232;