+++ Source/Shaders/Builtin/Constants/epsilon1.glsl
@@ -1,0 +1,7 @@
+/**
+ * 0.1
+ *
+ * @name czm_epsilon1
+ * @glslConstant
+ */
+const float czm_epsilon1 = 0.1;