+++ Source/Widgets/Viewer/viewerDynamicObjectMixin.js
@@ -4,6 +4,7 @@ define([
         '../../Core/defined',
         '../../Core/DeveloperError',
         '../../Core/defineProperties',
+        '../../Core/Event',
         '../../Core/EventHelper',
         '../../Core/ScreenSpaceEventType',
         '../../Core/wrapFunction',
@@ -14,6 +15,7 @@ define([
         defined,
         DeveloperError,
         defineProperties,
+        Event,
         EventHelper,
         ScreenSpaceEventType,
         wrapFunction,
@@ -48,8 +50,12 @@ define([
         if (viewer.hasOwnProperty('trackedObject')) {
             throw new DeveloperError('trackedObject is already defined by another mixin.');
         }
+        if (viewer.hasOwnProperty('onObjectTracked')) {
+            throw new DeveloperError('onObjectTracked is already defined by another mixin.');
+        }
 
         var eventHelper = new EventHelper();
+        var onObjectTracked = new Event();
         var trackedObject;
         var dynamicObjectView;
 
@@ -102,10 +108,6 @@ define([
                     return trackedObject;
                 },
                 set : function(value) {
-                    if (trackedObject !== value) {
-                        trackedObject = value;
-                        dynamicObjectView = defined(value) ? new DynamicObjectView(value, viewer.scene, viewer.centralBody.getEllipsoid()) : undefined;
-                    }
                     var sceneMode = viewer.scene.getFrameState().mode;
 
                     if (sceneMode === SceneMode.COLUMBUS_VIEW || sceneMode === SceneMode.SCENE2D) {
@@ -115,6 +117,25 @@ define([
                     if (sceneMode === SceneMode.COLUMBUS_VIEW || sceneMode === SceneMode.SCENE3D) {
                         viewer.scene.getScreenSpaceCameraController().enableTilt = !defined(value);
                     }
+
+                    if (trackedObject !== value) {
+                        trackedObject = value;
+                        dynamicObjectView = defined(value) ? new DynamicObjectView(value, viewer.scene, viewer.centralBody.getEllipsoid()) : undefined;
+                        onObjectTracked.raiseEvent(viewer, value);
+                    }
+                }
+            },
+
+            /**
+             * Gets an event that will be raised when an object is tracked by the camera.  The event
+             * has two parameters: a reference to the viewer instance, and the newly tracked object.
+             *
+             * @memberof viewerDynamicObjectMixin.prototype
+             * @type {Event}
+             */
+            onObjectTracked : {
+                get : function() {
+                    return onObjectTracked;
                 }
             }
         });